Remove old and unuded classes
