--- win-linux/src/linux/csinglewindow.h
@@ -1,78 +1,0 @@
-/*
- * (c) Copyright Ascensio System SIA 2010-2019
- *
- * This program is a free software product. You can redistribute it and/or
- * modify it under the terms of the GNU Affero General Public License (AGPL)
- * version 3 as published by the Free Software Foundation. In accordance with
- * Section 7(a) of the GNU AGPL its Section 15 shall be amended to the effect
- * that Ascensio System SIA expressly excludes the warranty of non-infringement
- * of any third-party rights.
- *
- * This program is distributed WITHOUT ANY WARRANTY; without even the implied
- * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR  PURPOSE. For
- * details, see the GNU AGPL at: http://www.gnu.org/licenses/agpl-3.0.html
- *
- * You can contact Ascensio System SIA at 20A-12 Ernesta Birznieka-Upisha
- * street, Riga, Latvia, EU, LV-1050.
- *
- * The  interactive user interfaces in modified source and object code versions
- * of the Program must display Appropriate Legal Notices, as required under
- * Section 5 of the GNU AGPL version 3.
- *
- * Pursuant to Section 7(b) of the License you must retain the original Product
- * logo when distributing the program. Pursuant to Section 7(e) we decline to
- * grant you any rights under trademark law for use of our trademarks.
- *
- * All the Product's GUI elements, including illustrations and icon sets, as
- * well as technical writing content are licensed under the terms of the
- * Creative Commons Attribution-ShareAlike 4.0 International. See the License
- * terms at http://creativecommons.org/licenses/by-sa/4.0/legalcode
- *
-*/
-
-#ifndef CSINGLEWINDOW_H
-#define CSINGLEWINDOW_H
-
-#include <QMainWindow>
-#include "cmainpanel.h"
-#include "cx11decoration.h"
-#include <QLabel>
-
-class CSingleWindow : public QMainWindow, public CX11Decoration
-{
-    Q_OBJECT
-public:
-    explicit CSingleWindow(const QRect& rect, const QString& title, QWidget * view);
-
-    bool holdView(int id) const;
-    QWidget * createMainPanel(bool, const QString&, QWidget *);
-    void setScreenScalingFactor(double factor);
-    void applyTheme(const std::wstring& themeid);
-
-    const QWidget * handle() const;
-
-private:
-    bool event(QEvent *event);
-    void resizeEvent(QResizeEvent *);
-    void mouseMoveEvent(QMouseEvent *);
-    void mousePressEvent(QMouseEvent *);
-    void mouseReleaseEvent(QMouseEvent *);
-    void mouseDoubleClickEvent(QMouseEvent *);
-
-private:
-    QWidget * m_pMainPanel;
-    QWidget * m_boxTitle = nullptr;
-    QPushButton * m_btnMaximize = nullptr,
-                * m_btnMinimize = nullptr,
-                * m_btnClose = nullptr;
-    QLabel * m_labelTitle = nullptr;
-
-    double m_dpiRatio;
-
-signals:
-
-public slots:
-    void pushButtonCloseClicked();
-};
-
-#endif // CSINGLEWINDOW_H
